Move to separate Analyzer
