+++ Philips.CodeAnalysis.MaintainabilityAnalyzers/Documentation/DocumentThrownExceptionsAnalyzer.cs
@@ -24,27 +24,12 @@ namespace Philips.CodeAnalysis.MaintainabilityAnalyzers.Documentation
 		private const string InformationalTitle = @"Throw only informational exceptions";
 		private const string InformationalMessageFormat = @"Specify context to the {0} exception, by using a constructor overload that sets the Message property.";
 		private const string InformationalDescription = @"Specify context to a thrown exception, by using a constructor overload that sets the Message property.";
-		private const string LocationsTitle = @"Avoid throwing exceptions from unexpected locations";
-		private const string LocationsMessageFormat = @"Avoid throwing exceptions from {0}, as this location is unexpected.";
-		private const string LocationsDescription = @"Avoid throwing exceptions from unexpected locations, like Finalizers, Dispose, Static Constructors, etc...";
 		private const string Category = Categories.Documentation;
 
 		private static readonly DiagnosticDescriptor DocumentRule = new(Helper.ToDiagnosticId(DiagnosticIds.DocumentThrownExceptions), DocumentTitle, DocumentMessageFormat, Category, DiagnosticSeverity.Error, isEnabledByDefault: false, description: DocumentDescription);
 		private static readonly DiagnosticDescriptor InformationalRule = new(Helper.ToDiagnosticId(DiagnosticIds.ThrowInformationalExceptions), InformationalTitle, InformationalMessageFormat, Category, DiagnosticSeverity.Error, isEnabledByDefault: false, description: InformationalDescription);
-		private static readonly DiagnosticDescriptor LocationsRule = new(Helper.ToDiagnosticId(DiagnosticIds.AvoidExceptionsFromUnexpectedLocations), LocationsTitle, LocationsMessageFormat, Category, DiagnosticSeverity.Error, isEnabledByDefault: false, description: LocationsDescription);
 
-		[SuppressMessage("", "IDE0090", Justification = "When type is removed the literal type cannot be deduced")]
-		private static readonly Dictionary<string, string> specialMethods = new Dictionary<string, string>
-		{
-			{ ".ctor", "constructor" },
-			{ ".cctor", "static constructor" },
-			{ "Equals", "Equals method" },
-			{ "GetHashCode", "GetHashCode method" },
-			{ "Dispose", "Dispose method" },
-			{ "ToString", "ToString method" }
-		};
-
-		public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(DocumentRule, InformationalRule, LocationsRule);
+		public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => ImmutableArray.Create(DocumentRule, InformationalRule);
 
 		public override void Initialize(AnalysisContext context)
 		{
@@ -92,46 +77,6 @@ namespace Philips.CodeAnalysis.MaintainabilityAnalyzers.Documentation
 				{
 					return;
 				}
-			}
-
-			// Check if the Parent is an unexpected location.
-			if (methodDeclaration is MethodDeclarationSyntax method)
-			{
-				if (specialMethods.TryGetValue(method.Identifier.Text, out string specialMethodKind))
-				{
-					var loc = method.Identifier.GetLocation();
-					Diagnostic diagnostic = Diagnostic.Create(LocationsRule, loc, specialMethodKind);
-					context.ReportDiagnostic(diagnostic);
-				}
-			}
-			else if (methodDeclaration is OperatorDeclarationSyntax { OperatorToken.Text: "==" or "!=" } operatorDeclaration)
-			{
-				var loc = operatorDeclaration.OperatorKeyword.GetLocation();
-				Diagnostic diagnostic = Diagnostic.Create(LocationsRule, loc, "equality comparison operator");
-				context.ReportDiagnostic(diagnostic);
-			}
-			else if (methodDeclaration is ConversionOperatorDeclarationSyntax { ImplicitOrExplicitKeyword.Text: "implicit" } conversionDeclaration)
-			{
-				var loc = conversionDeclaration.OperatorKeyword.GetLocation();
-				Diagnostic diagnostic = Diagnostic.Create(LocationsRule, loc, "implicit cast operator");
-				context.ReportDiagnostic(diagnostic);
-			}
-			else if(methodDeclaration is ConstructorDeclarationSyntax constructorDeclaration)
-			{
-				bool? withinExceptionClass =
-					(methodDeclaration.Parent as TypeDeclarationSyntax)?.Identifier.Text.EndsWith("Exception");
-				if ((withinExceptionClass.HasValue && (bool)withinExceptionClass) || constructorDeclaration.Modifiers.Any(SyntaxKind.StaticKeyword))
-				{
-					var loc = constructorDeclaration.Identifier.GetLocation();
-					Diagnostic diagnostic = Diagnostic.Create(LocationsRule, loc, "implicit cast operator");
-					context.ReportDiagnostic(diagnostic);
-				}
-			}
-			else if(methodDeclaration is DestructorDeclarationSyntax destructorDeclaration)
-			{
-				var loc = destructorDeclaration.Identifier.GetLocation();
-				Diagnostic diagnostic = Diagnostic.Create(LocationsRule, loc, "finalizer");
-				context.ReportDiagnostic(diagnostic);
 			}
 
 			var mentionedExceptions = methodDeclaration.GetLeadingTrivia()
